get into a buildable state
